+++ CHANGELOG.md
@@ -3,6 +3,15 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [1.3.0]
+### Added
+- Allow function locationClassName prop of SVGMap, CheckboxSVGMap and RadioSVGMap components
+
+### Changed
+- Update examples
+- Externalize Jest config
+- Update Jest config
+
 ## [1.2.0]
 ### Added
 - Create CheckboxSVGMap and RadioSVGMap components